Implement alternate location in %AppData% for the configuration file
When reading the configuration, the order of preference is:
- %AppData%/NegativeScreen/negativescreen.conf
- negativescreen.conf
- default embedded configuration

When editing the configuration, the path used for reading is used in priority.
Otherwise, the order of preference is reversed.
(That is, if no configuration is present in AppData, one will be created
in the working directory first, and AppData will only be used if that fails)